# .history/backend/storage/api-docs/api-docs_20260707034108.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HiKala API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "List Products",
                "description": "Get all products with pagination.",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "responses": {
                    "200": {
                        "description": "Products retrieved successfully."
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create Product",
                "description": "Create a new product with variants and attributes.",
                "operationId": "785644b7d7dd38af4b51dafd0fbc1b6d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "brand_id",
                                    "title",
                                    "slug"
                                ],
                                "properties": {
                                    "brand_id": {
                                        "description": "شناسه برند",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "description": "عنوان محصول",
                                        "type": "string",
                                        "example": "iPhone 15 Pro"
                                    },
                                    "slug": {
                                        "description": "اسلاگ محصول (آدرس سئویی)",
                                        "type": "string",
                                        "example": "iphone-15-pro"
                                    },
                                    "short_description": {
                                        "description": "توضیح مختصر",
                                        "type": "string",
                                        "example": "گوشی آیفون ۱۵ پرو با پردازنده A16"
                                    },
                                    "description": {
                                        "description": "توضیح کامل (HTML)",
                                        "type": "string",
                                        "example": "<p>توضیحات کامل محصول</p>"
                                    },
                                    "status": {
                                        "description": "وضعیت (active/inactive/draft)",
                                        "type": "string",
                                        "example": "active"
                                    },
                                    "meta_title": {
                                        "description": "عنوان سئو",
                                        "type": "string",
                                        "example": "iPhone 15 Pro"
                                    },
                                    "meta_keywords": {
                                        "description": "کلمات کلیدی سئو",
                                        "type": "string",
                                        "example": "iphone, apple"
                                    },
                                    "meta_description": {
                                        "description": "توضیحات سئو",
                                        "type": "string",
                                        "example": "خرید آیفون ۱۵ پرو"
                                    },
                                    "is_active": {
                                        "description": "فعال/غیرفعال",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "categories": {
                                        "description": "آرایه شناسه دسته‌بندی‌ها",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2
                                        ]
                                    },
                                    "variants": {
                                        "description": "لیست تنوع‌های محصول",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "sku": {
                                                    "description": "کد انبار",
                                                    "type": "string",
                                                    "example": "SKU-001"
                                                },
                                                "barcode": {
                                                    "description": "بارکد",
                                                    "type": "string",
                                                    "example": "1234567890123"
                                                },
                                                "price": {
                                                    "description": "قیمت اصلی (تومان)",
                                                    "type": "integer",
                                                    "example": 30000000
                                                },
                                                "sale_price": {
                                                    "description": "قیمت تخفیف‌خورده (تومان)",
                                                    "type": "integer",
                                                    "example": 25000000
                                                },
                                                "stock": {
                                                    "description": "موجودی انبار",
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "weight": {
                                                    "description": "وزن (گرم)",
                                                    "type": "integer",
                                                    "example": 200
                                                },
                                                "is_active": {
                                                    "description": "فعال/غیرفعال",
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "attributes": {
                                                    "description": "ویژگی‌های این تنوع",
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "attribute_value_id": {
                                                                "description": "شناسه مقدار ویژگی",
                                                                "type": "integer",
                                                                "example": 1
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully."
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{product}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Show Product",
                "description": "Get product details with variants and attributes.",
                "operationId": "db00b523e69d5dbc09c6edad76374bdc",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product retrieved successfully."
                    },
                    "404": {
                        "description": "Product not found."
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update Product",
                "description": "Update product information with variants.",
                "operationId": "f47cf55f4f3e0b134b5e7d9a8373725d",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "brand_id": {
                                        "description": "شناسه برند",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "description": "عنوان محصول",
                                        "type": "string",
                                        "example": "iPhone 15 Pro"
                                    },
                                    "slug": {
                                        "description": "اسلاگ محصول",
                                        "type": "string",
                                        "example": "iphone-15-pro"
                                    },
                                    "short_description": {
                                        "description": "توضیح مختصر",
                                        "type": "string",
                                        "example": "گوشی آیفون ۱۵ پرو"
                                    },
                                    "description": {
                                        "description": "توضیح کامل",
                                        "type": "string",
                                        "example": "<p>توضیحات کامل</p>"
                                    },
                                    "status": {
                                        "description": "وضعیت",
                                        "type": "string",
                                        "example": "active"
                                    },
                                    "meta_title": {
                                        "description": "عنوان سئو",
                                        "type": "string",
                                        "example": "iPhone 15 Pro"
                                    },
                                    "meta_keywords": {
                                        "description": "کلمات کلیدی",
                                        "type": "string",
                                        "example": "iphone, apple"
                                    },
                                    "meta_description": {
                                        "description": "توضیحات سئو",
                                        "type": "string",
                                        "example": "خرید آیفون ۱۵ پرو"
                                    },
                                    "is_active": {
                                        "description": "فعال/غیرفعال",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "categories": {
                                        "description": "آرایه شناسه دسته‌بندی‌ها",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2
                                        ]
                                    },
                                    "variants": {
                                        "description": "لیست تنوع‌های محصول",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "description": "شناسه تنوع (برای آپدیت)",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "sku": {
                                                    "description": "کد انبار",
                                                    "type": "string",
                                                    "example": "SKU-001"
                                                },
                                                "barcode": {
                                                    "description": "بارکد",
                                                    "type": "string",
                                                    "example": "1234567890123"
                                                },
                                                "price": {
                                                    "description": "قیمت اصلی",
                                                    "type": "integer",
                                                    "example": 30000000
                                                },
                                                "sale_price": {
                                                    "description": "قیمت تخفیف‌خورده",
                                                    "type": "integer",
                                                    "example": 25000000
                                                },
                                                "stock": {
                                                    "description": "موجودی",
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "weight": {
                                                    "description": "وزن",
                                                    "type": "integer",
                                                    "example": 200
                                                },
                                                "is_active": {
                                                    "description": "فعال/غیرفعال",
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "attributes": {
                                                    "description": "ویژگی‌های تنوع",
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "attribute_value_id": {
                                                                "description": "شناسه مقدار ویژگی",
                                                                "type": "integer",
                                                                "example": 1
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully."
                    },
                    "404": {
                        "description": "Product not found."
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete Product",
                "description": "Delete a product.",
                "operationId": "ac878db81fdb1333b21fc79e574a5a17",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted successfully."
                    },
                    "404": {
                        "description": "Product not found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}